@@ -3,7 +3,7 @@
     "description": "Kimai extension to configure a language, which will be used to format money, date and time values in invoices",
     "homepage": "https://www.kimai.org/store/invoice-format-fixation-bundle.html",
     "type": "kimai-plugin",
-    "version": "1.1",
+    "version": "2.0",
     "keywords": [
         "kimai",
         "kimai-plugin"
@@ -12,7 +12,6 @@
     "authors": [
         {
             "name": "Kevin Papst",
-            "email": "info@keleo.de",
             "homepage": "https://www.keleo.de"
         }
     ],
@@ -24,10 +23,11 @@
     "config": {
         "allow-plugins": {
             "composer/package-versions-deprecated": false,
-            "symfony/flex": false
+            "symfony/flex": false,
+            "symfony/runtime": false
         },
         "platform": {
-            "php": "7.3"
+            "php": "8.1.3"
         },
         "preferred-install": {
             "*": "dist"
@@ -36,7 +36,7 @@
     },
     "extra": {
         "kimai": {
-            "require": "1.17",
+            "require": 20000,
             "name": "Invoice format fixation"
         }
     },
@@ -55,8 +55,8 @@
         "friendsofphp/php-cs-fixer": "^3.0",
         "phpstan/phpstan": "^1.0",
         "phpstan/phpstan-symfony": "^1.0",
-        "kevinpapst/kimai2": "dev-master",
-        "symfony/console": "^4.0",
-        "symfony/event-dispatcher": "^4.0"
+        "kimai/kimai": "2.0.x-dev",
+        "symfony/console": "^6.0",
+        "symfony/event-dispatcher": "^6.0"
     }
 }
