@@ -1,7 +1,7 @@
 <?php
 
 /*
- * This file is part of the "Invoice format fixation" plugin for Kimai 2.
+ * This file is part of the "Invoice format fixation bundle" for Kimai.
  * All rights reserved by Kevin Papst (www.kevinpapst.de).
  *
  * For the full copyright and license information, please view the LICENSE
@@ -11,18 +11,12 @@
 namespace KimaiPlugin\InvoiceFormatFixationBundle\DependencyInjection;
 
 use App\Plugin\AbstractPluginExtension;
-use Exception;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader;
 
 class InvoiceFormatFixationExtension extends AbstractPluginExtension
 {
-    /**
-     * @param array $configs
-     * @param ContainerBuilder $container
-     * @throws Exception
-     */
     public function load(array $configs, ContainerBuilder $container): void
     {
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
